+++ src/com/three/psyco/model/dao/MemberDAOImpl.java
@@ -1,11 +1,13 @@
 package com.three.psyco.model.dao;
 
+import java.sql.SQLException;
 import java.util.HashMap;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.stereotype.Repository;
 
 import com.three.psyco.model.dto.MemberDTO;
+import com.three.psyco.model.dto.MemberShopDTO;
 
 @Repository
 public class MemberDAOImpl implements MemberDAO {
@@ -82,5 +84,21 @@ public class MemberDAOImpl implements MemberDAO {
 	public int modifyNormalUserPro(MemberDTO dto) {
 		int result = sqlSession.update("member.modifyNormalUserPro", dto);
 		return result;
+	}
+	
+	@Override
+	public void insertMemberShop(MemberShopDTO dto) throws SQLException {
+		System.out.println(dto.getShop_name());
+		System.out.println(dto.getShop_num());
+		System.out.println(dto.getShop_phone());
+		System.out.println(dto.getStatus());
+		System.out.println("shop 이미지: " + dto.getShop_img());
+		System.out.println("takeout"+dto.getTakeout());
+		if(dto.getTakeout() == null) {
+			dto.setTakeout("0");
+			sqlSession.insert("member.insertMemberShop",dto);	
+		}else {
+			sqlSession.insert("member.insertMemberShop",dto);	
+		}
 	}
 }
