+++ src/sql/memberSQL.xml
@@ -44,4 +44,8 @@
   	<update id="modifyNormalUserPro" parameterType="memberDTO">
   		update member set pw=#{pw}, name=#{name}, birth=#{birth}, phoneNum=#{phoneNum}, gender=#{gender}, nickname=#{nickname}, email=#{email} where member_Num=#{member_Num} and member_Id=#{member_Id}
   	</update>
+  	
+  	<insert id="insertMemberShop" parameterType="MemberShopDTO">
+  		insert into shop values(#{shop_num},#{shop_name},#{shop_phone},#{operating_time},#{address},#{origin},#{takeout},#{shop_img},#{license_number},sysdate,sysdate,#{status},#{approve_status},#{member_num})
+  	</insert>
   </mapper>
